+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.0.1 ] - 2024-06-22
+## [ 1.1.0 ] - 2024-06-23
+### Added
+- IProcessingProgress.resetProcessingStatusMessage to reset status messages.
+- IProcessingStatistic.isEmpty method.
+- ProcessingUnitRunnable.toString method to print proper the current status.
+- Added test cases.
+
+### Fixed
+- Throttling issue after suspend and resume.
+- Several null checks.
+
+### Changed
+- Improved logging.
 
 ## [ 1.0.0 ] - 2024-06-22
 ### Changed